+++ Changelog.md
@@ -8,7 +8,7 @@ OpenCore Changelog
 - Fixed microcode reading on Intel CPUs
 - Fixed SMBIOS Type4 External Clock values
 - Improved Windows compatibility on some setups (acidanthera/bugtracker#614)
-- Added `SupportsCsm` and `ReplaceableMemory` options in `PlatformInfo/Generic`
+- Added `SupportsCsm` and option in `PlatformInfo/Generic`
 
 #### v0.5.3
 - Update builtin firmware versions